better aptitude and synaptic detection
This is not tested, I haven't used Gambas in some time and there was a difference in the 'IF' statements I think. Also, the binary needs to be rebuilded for this changes to take affect but I can't do that, I'm running custom distribution and Gambas (2 and 3, esecially 3) are pain in the rear to compile because they both need QT toolkit and WebKit installed and I don't need them.